HashFolia/requirements.txt
Pillow
Flask
Vizhash

